requirements.txt
# Django
Django==5.1
django-environ==0.11.2
djangorestframework==3.15.2
django-filter==24.3

# Transformers/Machine Translation
transformers[torch]
torch==2.2.1
sentencepiece==0.2.0
sacremoses==0.1.1
